Fix - CRUD de citas oki doki

+++ taller-be/Controllers/UsuariosController.cs
@@ -154,7 +154,7 @@ namespace taller_be.Controllers
 
 
         // GET: Usuarios/Delete/5
-        public async Task<IActionResult> Delete(decimal? id)
+        public async Task<IActionResult> Delete(int? id)
         {
             if (id == null || _context.Usuarios == null)
             {
@@ -174,7 +174,7 @@ namespace taller_be.Controllers
         // POST: Usuarios/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(decimal id)
+        public async Task<IActionResult> DeleteConfirmed(int id)
         {
             if (_context.Usuarios == null)
             {
